Move authOptions out of route.ts
