+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,19 +1,6 @@
 
 import NextAuth from 'next-auth'
-import GithubProvider from 'next-auth/providers/github'
-import prisma from '@/lib/prisma/prisma'
-import { PrismaAdapter } from '@auth/prisma-adapter'
-import { Adapter } from 'next-auth/adapters'
-
-export const authOptions = {
-  adapter: PrismaAdapter(prisma) as Adapter,
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID!,
-      clientSecret: process.env.GITHUB_CLIENT_SECRET!
-    })
-  ]
-}
+import { authOptions } from '@/lib/oauth/authOptions'
 
 const handler = NextAuth(authOptions)
