+++ src/lib/oauth/authOptions.ts
@@ -1,0 +1,14 @@
+import GithubProvider from 'next-auth/providers/github'
+import prisma from '@/lib/prisma/prisma'
+import { PrismaAdapter } from '@auth/prisma-adapter'
+import { Adapter } from 'next-auth/adapters'
+
+export const authOptions = {
+  adapter: PrismaAdapter(prisma) as Adapter,
+  providers: [
+    GithubProvider({
+      clientId: process.env.GITHUB_CLIENT_ID!,
+      clientSecret: process.env.GITHUB_CLIENT_SECRET!
+    })
+  ]
+}
